+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Unreleased
-[no unreleased changes yet]
+- Stop creating a new alpha version after release.
 
 ## v2.0.1 (2025-03-19)
 - Add `rake` as a dependency.